Fixing example warnings.

+++ examples/gmres/util.h
@@ -41,9 +41,9 @@
 
 inline void printMatrix (DenseMatrix &M, const char *name) {
     printf("Matrix %s:\n", name);
-    for (int row = 0; row < M.rows(); row++) {
-        printf("row %2d: ", row + 1);
-        for (int col = 0; col < M.cols(); col++)
+    for (size_t row = 0; row < M.rows(); row++) {
+        printf("row %2d: ", (int)row + 1);
+        for (size_t col = 0; col < M.cols(); col++)
             printf("%6f ", M(row, col));
         printf("\n");
     }
